+++ src/nfc/read_only_service.hpp
@@ -36,6 +36,7 @@ enum class ReadState {
   stabilizing,
   reading,
   openprinttag,
+  blank,
   unsupported,
   multiple,
   error